requirements.txt
greenlet==3.2.3
mysql-connector-python==9.3.0
numpy==2.3.1
pandas==2.3.1
pymysql==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlalchemy==2.0.41
typing-extensions==4.14.1
tzdata==2025.2
